Fixed the analysis tools to work with the new dataloader and the parser.
