+++ bone_remodelling/analysis/cli.py
@@ -13,15 +13,8 @@ def cli(config: ConfigurationParameters, argv: list[str]) -> None:
         "--type",
         type=str,
         default="max_force_sweep",
-        choices=["forward_parameter_sweep", "max_force_sweep", "visualise_forces", "diversity_metrics"]
+        choices=["forward_parameter_sweep", "max_force_sweep", "visualise_forces", "diversity_metrics"],
         help="Analysis tools used for figure generation.",
-    )
-    parser.add_argument(
-        "--force_type",
-        type=str,
-        choices=["merger", "triangular"],
-        default="triangular",
-        help="Type of force profile to generate ('merger' or 'triangular').",
     )
     parser.add_argument(
         "--num_samples",
@@ -37,11 +30,14 @@ def cli(config: ConfigurationParameters, argv: list[str]) -> None:
         sweep_file = sweep_directory / Path("sweep_results.csv")
         forward_parameter_sweep(sweep_file, num_samples=args.num_samples)
     elif args.type == "max_force_sweep":
-        from bone_remodelling.analysis.max_force_sweep import run as max_force_sweep  # noqa: I001, PLC0415
-        max_force_sweep(config, force_type=args.force_type)
+        from bone_remodelling.analysis.max_force_sweep import run_sweep_and_analyse as max_force_sweep  # noqa: I001, PLC0415
+        max_force_sweep(config.output_dir / Path("sweeps", "max_force_sweep"))
     elif args.type == "visualise_forces":
-        from bone_remodelling.analysis.visualise_forces import run as visualise_forces  # noqa: I001, PLC0415
-        visualise_forces(config)
+        from bone_remodelling.analysis.visualise_force_profiles import analyse_force_generator as visualise_forces  # noqa: I001, PLC0415
+        visualise_forces()
+    elif args.type == "visualise_database":
+        from bone_remodelling.analysis.visualise_database import analyse_raw_data as visualise_database  # noqa: I001, PLC0415
+        visualise_database(config.output_dir / Path("raw"))
     elif args.type == "diversity_metrics":
         from bone_remodelling.analysis.diversity_metrics import run_diversity_metrics as diversity_metrics  # noqa: I001, PLC0415
         diversity_metrics(config.output_dir / Path("raw"))
