+++ bone_remodelling/surrogate_model/surrogate_parameters.py
@@ -38,7 +38,7 @@ class SurrogateTrainParameters(TrainParameters):
     model_path: Path
     device: torch.device
     epochs: int = 1_000
-    batch_size: int = 256
+    batch_size: int = 64
     learning_rate: float = 1e-3
     patience_lr_scheduler: int = 10
     factor_lr_scheduler: float = 0.5
